Replace Conv2dGw8.match argument "conv2d" with "module".
Modify Conv2dGw8.match to ensure the module is not already a Conv2dGw8 module.

Modify run_layer_test to use spio transform to create the spio model.

Add private function of spio.transform._transform._transform that performs the spio module replacement and also returns the number of replacements.

Implement spio.transform._transform.transform to call the private function and ignore the num_replacements return value.

Modify run_layer_test to check that exactly one spio module was matched.

Moved the get_full_name_with_underscores and get_full_name functions to spio.util

Add spio logging to performance_model_cache and transform.

Add torchcompile and torchcompile_mode arguments to run_layer_test()

Modify run_layer_test() to use torch.autocast

Modify test_conv2d_gw8_layer to test the layer with and without torchcompile.

Remove type checks from Conv2dGw8Params.from_tensors.

Fix placement of torch.compile in profile_timm.py

Add option --torchcompile-mode to profile_timm.py

Remove the --torchcompile-mode=reduce-overhead tests from benchmark_model.bash.

Spio does not work with reduce-overhead mode, and PyTorch does not benefit from it.

Modify conv2d_gw8_wgrad_kernel to return grad_weight with float32 data-type.

Also require user to initialize grad_weight to zeros before calling the kernel.

Note that the kernel's reflection already specified Init.ZERO initialization, so tests and benchmarks were already doing this.

Change the Conv2dGw8WgradKernel reflection to indicate that grad_weight's dtype is now float32.

Modify conv2d_gw8_backward_op to initialize grad_weight to zeros with float32 dtype.

Also modify conv2d_gw8_backward_op to use float32 precision for the grad_bias calculation.

Modify conv2d_gw8_setup_context to explicitly convert input, weight, and bias to torch.float16.

Note that the benchmark results are the same after these changes.

Reinvigorate the profile_fiesta.py script.

Use an inverted residual block to benchmark grouped convolutions in situ.

Add ConvFirst block option to profile_fiesta.py

Add batch-size sweep option to profile_fiest.py.

Add script to plot the kernel benchmark results.

Add effective bandwidth plot to plot_kernel_bench.py

Add extra MLP blocks to profile_fiesta benchmarks.

Add extra parameters to the benchmark directory name.

Add support for MBConv block benchmark plots.

Add parameter details to the plot filename.

Add torch kernels to plots.

Add double-underscores before and after device-name in the benchmark directory name.

Add suptitle to plots with layer params and device name.

Add the device name to the plot file name.

Add assertions to ensure that all kernels are matched.

Add assertions to ensure that the spio and torch data files use the same layer parameters.

Add assertion to ensure that both experiments use the same device and block.

Check that the experiment directory is not empty.

Remove redundant benchmark of last batch size.

Add support for plotting PyTorch depthwise convolution profiles.

Make group_width and stride fields of the Conv2dGw8Params dataclass.

Previously, these were constant class data members.

Adjust params_and_configs_to_data_frame to drop the "group_width" and "stride" params.

Add help message to benchmark_model.bash

Move effective bandwidth calculation into function add_eff_bandwidth_gb_s().

Add second plot with latency versus batch size.

Add new KernelParamsLogger class.

Modified the performance model cache so that it no longer compilers kernels.

It just returns the predicted best configuration.

Move kernel compilation and load to KernelCache.get().

Remove obsolete "config" argument from kernel getter methods.

Remove obsolete auto-tuning code path from KernelCache.

Replace KernelCache.get "args" argument with "device".

Update all calls to this function to pass the device.

Remove unused get_kernel methods from Kernel subclasses.

Add KernelCache overlay.

Add get_kernel_cache() accessors to  Kernel subclasses.

Enable kernel parameter logging in KernelCache.get

Add a "device" argument to  KernelParamsLogger.log_params.

Add KernelParamsLogger, log_kernel_params, and KernelParams to the exports of the kernel subpackage.

Add "results/" directory to .gitignore

Move functions preprocess_data_file from plot_kernel_bench.py to subpackage spio.data

Add first draft of --benchmark-configs option to profile_fiesta.py

Reduce number of samples used by benchmark_configs to minimum number of kernel configs across all benchmarked kernels.

This makes the sampling work without replacement in each list of kernel configs.

Add sanity check for empty Kernel cubins.

Add encoding of kernel parameters to the end of spio kernel names.

Continue to use "base name" without kernel parameters in the performance model name.

Add clear_cubin argument to method Kernel.unload with default value True.

Add compile_kernels function to compile a list of configured kernel objects in parallel.

This is a generalization of the compile_kernel_configs function which only works with a single kernel class.

Modify profile_fiesta.py to compile all kernels in parallel in the benchmark_configs function.

Modify compile_kernels() to return None.

Modify profile_fiesta benchmark_configs() to only compile the random selection of kernel configurations that will be profiled.

Move the ParseKwargs function from the benchmark.py script to spio.util

Add the --timm-model option to profile_fiest.py

Add the --timm-model-kwargs option too.

+++ spio/reflection/conv2d_gw8_reflection.py
@@ -101,7 +101,7 @@ def register_conv2d_gw8_reflections():
                 output=ArgInfo(dtype=torch.float16, output=True, init=Init.EMPTY),
                 grad_output=ArgInfo(dtype=torch.float16, grad_of="output"),
                 grad_weight=ArgInfo(
-                    dtype=torch.float16, init=Init.ZERO, grad_of="weight"
+                    dtype=torch.float32, init=Init.ZERO, grad_of="weight"
                 ),
             ),
             args=["grad_weight", "input", "grad_output"],
